Renamed ldaporm/typing.py to ldaporm/types.py to avoid an import problem in mypy itself
While trying to figure out how to get mypy to recognize CharField as str, I discovered that mypy was actually crashing when analyzing some files in the project, because it would attempt to import stuff from the standard library's `typing` module, but the `typing.py` file would cause the import to go astray and crash.

This rename fixes that problem, and updates the one local file that references it.