Removed basedpyright config from pyproject.toml
The existence of this stanza would break my own local basepyright configs for ALL of the projects I had open in the same workspace, because the plugin looks for *any* project that has such a stanza in its pyproject.toml, and *replaces* the settings in the IDE with the settings from pyproject.toml. But even worse, basepyright doesn't even have a python_version setting any more, so I don't think this was doing anything in the first place.

If the typeCheckingMode was doing something, you can replicate it by using folder-level Cursor settings (in a .vscode/settings.json file at the root of the repo) or workspace-level Cursor settings.