Somewhat iffy class change for CharListField
In order to make CharListFields play well with MyPy, I had to change the base field type to Field instead of CharField. I _think_ the other change I made, combined with the way the CharListField code already overrides everything from CharField, means this will still work, but I'm not 100% certain, so I put it in this separate commit.